made the website work

+++ spring boot project/ai-wrapper-java/src/main/java/com/example/aiplanner/dto/PlanRequest.java
@@ -2,15 +2,15 @@ package com.example.aiplanner.dto;
 
 public class PlanRequest {
     private String goal;
-    private Integer hoursPerWeek;
+    private Double hoursPerWeek;
     private Integer horizonDays;
     private String level;
 
     public String getGoal() { return goal; }
     public void setGoal(String goal) { this.goal = goal; }
 
-    public Integer getHoursPerWeek() { return hoursPerWeek; }
-    public void setHoursPerWeek(Integer hoursPerWeek) { this.hoursPerWeek = hoursPerWeek; }
+    public Double getHoursPerWeek() { return hoursPerWeek; }
+    public void setHoursPerWeek(Double hoursPerWeek) { this.hoursPerWeek = hoursPerWeek; }
 
     public Integer getHorizonDays() { return horizonDays; }
     public void setHorizonDays(Integer horizonDays) { this.horizonDays = horizonDays; }
